Outsourced tests for ivp examples to separate file
